Squashed commit of the following:
commit 4724395752efeb965f4076dec60de2c86fe8ff71
Author: danielSanchez <daniel.sanchez14@udea.edu.co>
Date:   Thu May 21 06:36:30 2026 -0500

    Use GeneradorProceduralSuavizado; tweak I/O & dask

    Introduce GeneradorProceduralSuavizado (new service) implementing smoothing rules: filters transitions to avoid infinite rests, limits large melodic jumps, and selects durations for generated events. main.py switched to this new generator, changed data_dir to the project data root, made MIDI discovery recursive (rglob), and updated the Dask client to use os.cpu_count() workers with one thread per worker. Also increased Dask bag partitions in ConstructorCorpusMidi from 8 to 64 to improve parallel processing of input files.

commit 34ddadd29304ba689803560a4071054b866884e9
Author: danielSanchez <daniel.sanchez14@udea.edu.co>
Date:   Wed May 20 23:01:25 2026 -0500

    Pass client to OrquestadorMapReduce and defer close

    Pass the Mongo client into OrquestadorMapReduce and avoid closing the client before MapReduce runs. Use the map-reduce result directly (removed transformar_datos_agrupados and orquestador.cerrar()), and close the client after saving the generated sequence file. Also add the generated MIDI output file output/secuencia_generada.mid.

commit 02ebd7721c122d8919d91013f26d264e66f9c22d
Merge: 70423fe b273dd4
Author: danielSanchez <daniel.sanchez14@udea.edu.co>
Date:   Wed May 20 22:20:12 2026 -0500

    Merge branch 'feature/matrix-generator-audio' into test/testing-dask

commit b273dd4c712379c964a8a87f9a8f178fb574352a
Author: danielSanchez <daniel.sanchez14@udea.edu.co>
Date:   Wed May 20 22:17:54 2026 -0500

    Accept external Client in OrquestadorMapReduce

    Change OrquestadorMapReduce to accept a Client via its constructor instead of instantiating one internally. __init__ now takes client: Client and assigns it to self._client (previously created with n_workers and dashboard_address). The cerrar() method that closed the client was removed. This enables dependency injection/reuse of an existing client (and easier testing) and prevents the orchestrator from unilaterally closing an externally managed client. The dashboard link print remains unchanged.

commit 70423fe7b894bbbb3f97607f5e3ae419e398dee8
Author: danielSanchez <daniel.sanchez14@udea.edu.co>
Date:   Wed May 20 22:15:45 2026 -0500

    Add Dask processing, generation and MIDI export

    Integrate Dask-based parallel MIDI ingestion and extend the pipeline to generate Markov matrix, procedural sequence and export to MIDI. main.py: switch data dir to data/2004, create Dask Client, batch-load files via ConstructorCorpusMidi.agregar_archivos, close client, run MapReduce, build Markov matrix, generate procedural sequence and export result (removed interactive prompt). ConstructorCorpusMidi.py: refactor to use dask.bag for parallel processing, add agregar_archivos, convert many instance helpers to static functions, simplify event extraction and standardize corpus/event attributes (_lista_eventos). ExportadorMidi.py: adapt to renamed attributes (check _lista_eventos and use evento._nota._nota_midi). Note: API callers should now provide lists of paths to agregar_archivos and expect the adjusted attribute names for sequence and note fields.

commit 5614c46a7fac16212605fc3af68a763ddb1cebea
Merge: 39132f6 6d7dc0a
Author: danielSanchez <daniel.sanchez14@udea.edu.co>
Date:   Wed May 20 20:45:36 2026 -0500

    Merge branch 'feature/midi-processor' into feature/matrix-generator-audio

commit 39132f652b60f9e6fd59193ec555f25f3ba9a2af
Author: danielSanchez <daniel.sanchez14@udea.edu.co>
Date:   Wed May 20 20:44:32 2026 -0500

    Use int durations; add Markov generator and tests

    Change duration/tiempos types from float to int across DTO/model classes to standardize durations as integer milliseconds (EstadisticasDeTransicion, EventoMusical, TransicionMarkov). Add services: GeneradorMatrizMarkov (builds a MatrizMarkov from grouped transition data) and GeneradorProcedural (generates SecuenciaMusical procedurally using Markov transitions and random selection of durations). Add tests (TestGeneradorProcedural.py, TestMatrixGeneration.py) with mocked transition data to exercise matrix generation and sequence generation. The new generator implementations rely on MatrizMarkov, NotaMusical, TransicionMarkov and DataTransicionAgrupada and implement probability calculation and random interval-based transition selection.

+++ model/TransicionMarkov.py
@@ -6,4 +6,4 @@ class TransicionMarkov:
     _nota_origen: NotaMusical
     _nota_destino: NotaMusical
     _probabilidad_transicion: float
-    _lista_tiempos: list[float]
+    _lista_tiempos: list[int]
